requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
langchain==0.2.16
langchain-openai==0.1.25
langgraph==0.2.28
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
openai==1.51.0
